Extract one_way_shared_memory to IPC and prepare libdd-remote-config for python
