+++ datadog-sidecar/src/service/telemetry.rs
@@ -8,10 +8,10 @@ use std::sync::OnceLock;
 use tokio::sync::mpsc;
 use tracing::{debug, info, warn};
 
-use crate::one_way_shared_memory::OneWayShmWriter;
 use crate::primary_sidecar_identifier;
 use base64::prelude::BASE64_URL_SAFE_NO_PAD;
 use base64::Engine;
+use datadog_ipc::one_way_shared_memory::OneWayShmWriter;
 use datadog_ipc::platform::NamedShmHandle;
 use std::collections::{HashMap, HashSet, VecDeque};
 use std::ffi::CString;
