+++ libdd-remote-config/Cargo.toml
@@ -27,6 +27,8 @@ client = [
 ]
 
 regex-lite = ["libdd-common/regex-lite"]
+# Expose select types (e.g. RemoteConfigCapabilities) to Python via PyO3.
+pyo3 = ["dep:pyo3"]
 # Enable HTTPS support in the fetcher (rustls + ring via libdd-common).
 https = ["libdd-common/https"]
 # FIPS-compliant crypto provider (aws-lc-rs via libdd-common). Unix only.
@@ -38,6 +40,7 @@ anyhow = { version = "1.0" }
 libdd-common = { path = "../libdd-common", version = "4.2.0", default-features = false }
 libdd-trace-protobuf = { path = "../libdd-trace-protobuf", version = "3.0.2", optional = true }
 hyper = { workspace = true, optional = true, default-features = false }
+pyo3 = { version = "0.28", optional = true, default-features = false, features = ["macros"] }
 http-body-util = {version = "0.1", optional = true }
 http = { version = "1.1", optional = true }
 base64 = { version = "0.22.1", optional = true }
