+++ libdd-remote-config/src/file_change_tracker.rs
@@ -57,12 +57,14 @@ impl<S: FilePath> ChangeTracker<S> {
         let files = HashSet::from_iter(files.into_iter().map(FilePathBasedArc));
         let mut changes = vec![];
 
-        for file in files.difference(&self.last_files) {
-            changes.push(Change::Add(file.0.clone()));
-        }
-
+        // Emit removals before additions/updates so that consumers observe
+        // configuration deletions strictly before new or updated configs.
         for file in self.last_files.difference(&files) {
             changes.push(Change::Remove(file.0.clone()));
+        }
+
+        for file in files.difference(&self.last_files) {
+            changes.push(Change::Add(file.0.clone()));
         }
 
         for (updated_file, old_contents) in updated.into_iter() {
@@ -74,5 +76,51 @@ impl<S: FilePath> ChangeTracker<S> {
 
         self.last_files = files;
         changes
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::{RemoteConfigPath, RemoteConfigProduct, RemoteConfigSource};
+
+    struct TestFile(RemoteConfigPath);
+
+    impl FilePath for TestFile {
+        fn path(&self) -> &RemoteConfigPath {
+            &self.0
+        }
+    }
+
+    fn file(name: &str) -> Arc<TestFile> {
+        Arc::new(TestFile(RemoteConfigPath {
+            source: RemoteConfigSource::Employee,
+            product: RemoteConfigProduct::ApmTracing,
+            config_id: "id".to_string(),
+            name: name.to_string(),
+        }))
+    }
+
+    #[test]
+    fn removals_are_emitted_before_additions() {
+        let mut tracker: ChangeTracker<TestFile> = ChangeTracker::default();
+        let a = file("a");
+        // First round: "a" is added.
+        tracker.get_changes::<()>(vec![a.clone()], vec![]);
+
+        // Second round: "a" is gone, "b" appears. The removal of "a" must be
+        // emitted before the addition of "b".
+        let b = file("b");
+        let changes = tracker.get_changes::<()>(vec![b.clone()], vec![]);
+
+        assert_eq!(changes.len(), 2);
+        match &changes[0] {
+            Change::Remove(f) => assert_eq!(f.path().name, "a"),
+            _ => panic!("expected the removal to be emitted first"),
+        }
+        match &changes[1] {
+            Change::Add(f) => assert_eq!(f.path().name, "b"),
+            _ => panic!("expected the addition to be emitted after the removal"),
+        }
     }
 }
